jnj-orca-server-sprint05/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "Production-ready FastAPI template"
authors = ["Fission Labs"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
pydantic-settings = "^2.4.0"
sqlmodel = "^0.0.24"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
fastapi-pagination = "^0.14.1"
sqlakeyset = "^2.0.1746777265"
fastapi-filter = "^2.0.1"
pandas = "^2.3.2"
pypdf2 = "^3.0.1"
spire-doc = "^13.8.0"
pillow = "^11.3.0"
python-pptx = "^1.0.2"
pypdf = "^6.1.0"
reportlab = "^4.4.4"
celery = "^5.5.3"
zipstream = "^1.1.4"
redis = "^6.4.0"
boolean-py = "^5.0"
boto3 = "^1.40.37"
openpyxl = "^3.1.5"
aiofiles = "^24.1.0"
ldap3 = "^2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
alembic = "^1.16.5"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
